remove results page duplications
