+++ src/Defra.Trade.ReMoS.AssuranceService.UI.Hosting/Pages/SelfServe/Establishments/PostcodeSearch.cshtml.cs
@@ -1,12 +1,12 @@
 using Defra.Trade.ReMoS.AssuranceService.UI.Core.Configuration;
 using Defra.Trade.ReMoS.AssuranceService.UI.Core.Interfaces;
-using Defra.Trade.ReMoS.AssuranceService.UI.Hosting.Pages.Shared.Establishments;
+using SignUp = Defra.Trade.ReMoS.AssuranceService.UI.Hosting.Pages.Establishments;
 using Microsoft.FeatureManagement.Mvc;
 
 namespace Defra.Trade.ReMoS.AssuranceService.UI.Hosting.Pages.SelfServe;
 
 [FeatureGate(FeatureFlags.SelfServeMvpPlus)]
-public class PostcodeSearchModel : PostcodeSearchModelBase
+public class PostcodeSearchModel : SignUp.PostcodeSearchModel
 {
     public PostcodeSearchModel(ILogger<PostcodeSearchModel> logger, ITraderService traderService) : base(logger, traderService)
     { }
